reorganized the structure of the folders for the components
