+++ src/components/pages/Projects.jsx
@@ -1,13 +1,11 @@
 // import { useState } from "react";
-import { ProjectCard } from "./UI/Card";
+import { ProjectCard } from "../UI/Card";
 
 import { Grid2 as Grid } from "@mui/material";
 
-
-import { projectData } from "../assets/js/projects";
+import { projectData } from "../../assets/js/projects";
 
 export const Projects = () => {
-  //   const [project, setProject] = useState([]);
 
   return (
     <Grid
@@ -18,15 +16,12 @@ export const Projects = () => {
         minHeight: "50vh",
         paddingTop: "50px", // Padding around content
         paddingBottom: "100px", // Adjust this value to ensure there's space for the footer
-        background:"#B7B7B7",
-        paddingLeft:"20px",
-        paddingRight:"20px"
+        background: "#B7B7B7",
+        paddingLeft: "20px",
+        paddingRight: "20px",
       }}
     >
       <ProjectCard data={projectData}></ProjectCard>
-      {/* <ProjectCard></ProjectCard>
-      <ProjectCard></ProjectCard>
-      <ProjectCard></ProjectCard> */}
     </Grid>
   );
 };
